# Cybernetic-openapi3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cybernetic",
    "description": "App Sec project",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/products/uploads": {
      "post": {
        "tags": [
          "Customer"
        ],
        "summary": "Upload Product Image",
        "description": "Try Update Product Image as User \r\nUsername : Tester\r\nEmail : test@test.com",
        "operationId": "UploadProductImage",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "encoding": {},
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/products": {
      "post": {
        "tags": [
          "Customer"
        ],
        "summary": "Add Product",
        "description": "Add A New Product\r\nUsername : Tester\r\nEmail : test@test.com",
        "operationId": "AddProduct",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProductRequest"
              },
              "example": {
                "retail_price": 135.0,
                "name": "Blackpods Pro2",
                "description": "This is a scam",
                "pic_filename": "scampods_3529963a5-a2fa-458c-8b71-b4f8acdc1e12.jpg",
                "stock": 30
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Retrieve Products",
        "operationId": "RetrieveProducts",
        "parameters": [
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 20
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/search/product": {
      "get": {
        "tags": [
          "Customer"
        ],
        "summary": "Search Product",
        "description": "Search for Product by Key",
        "operationId": "SearchProduct",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "Airpods 2';"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "Login into Cybernetic, returns JWT auth token",
        "operationId": "Login",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "test@test.com ' or 1=1 --",
                "password": "not_correct_password"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "Content-Length": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "227"
                  }
                }
              },
              "Server": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Werkzeug/1.0.1 Python/3.8.2"
                  }
                }
              },
              "Date": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Tue, 16 Jun 2020 10:45:15 GMT"
                  }
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Login-SuccessfulLogin"
                },
                "example": {
                  "success": true,
                  "message": "Successfully logged in.",
                  "auth_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1OTIzMDQzMjAsImlhdCI6MTU5MjMwNDMxNSwic3ViIjoxfQ.tOVFvLFLmZLefEtmRu2oqeut41_crOhZW8cGS67Rpns"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewAddress-Alreadyexist"
                },
                "example": {
                  "message": "Incorrect username or password",
                  "success": false
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register",
        "description": "Register a new user and returns JWT auth token of the corresponding user",
        "operationId": "Register",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "username": "tester2",
                "email": "test2@test.com",
                "password": "password"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "headers": {
              "Content-Length": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "224"
                  }
                }
              },
              "Server": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Werkzeug/1.0.1 Python/3.8.2"
                  }
                }
              },
              "Date": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Tue, 16 Jun 2020 10:43:18 GMT"
                  }
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Register-Successful"
                },
                "example": {
                  "auth_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1OTIzMDQyMDMsImlhdCI6MTU5MjMwNDE5OCwic3ViIjoxfQ.979DYzxQyQa1opjxBPNEsCnBHZs6uW2TlRjvC0skYAA",
                  "message": "Successfully registered.",
                  "success": true
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/auth/forget-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forget Password",
        "description": "Email a link to existing email for password reseting",
        "operationId": "ForgetPassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetPasswordRequest"
              },
              "example": {
                "email": "dylanliew0503@gmail.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/auth/forget-password/new-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "New Password",
        "operationId": "NewPassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPasswordRequest"
              },
              "example": {
                "user_id": "2",
                "code": "750248",
                "password": "newpassword"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "title": "LoginRequest",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "test@test.com ' or 1=1 --",
          "password": "not_correct_password"
        }
      },
      "Login-SuccessfulLogin": {
        "title": "Login-SuccessfulLogin",
        "required": [
          "success",
          "message",
          "auth_token"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "auth_token": {
            "type": "string"
          }
        },
        "example": {
          "success": true,
          "message": "Successfully logged in.",
          "auth_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1OTIzMDQzMjAsImlhdCI6MTU5MjMwNDMxNSwic3ViIjoxfQ.tOVFvLFLmZLefEtmRu2oqeut41_crOhZW8cGS67Rpns"
        }
      },
      "RegisterRequest": {
        "title": "RegisterRequest",
        "required": [
          "username",
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "username": "tester2",
          "email": "test2@test.com",
          "password": "password"
        }
      },
      "Register-Successful": {
        "title": "Register-Successful",
        "required": [
          "auth_token",
          "message",
          "success"
        ],
        "type": "object",
        "properties": {
          "auth_token": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "example": {
          "auth_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1OTIzMDQyMDMsImlhdCI6MTU5MjMwNDE5OCwic3ViIjoxfQ.979DYzxQyQa1opjxBPNEsCnBHZs6uW2TlRjvC0skYAA",
          "message": "Successfully registered.",
          "success": true
        }
      },
      "ForgetPasswordRequest": {
        "title": "ForgetPasswordRequest",
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "example": {
          "email": "dylanliew0503@gmail.com"
        }
      },
      "NewPasswordRequest": {
        "title": "NewPasswordRequest",
        "required": [
          "user_id",
          "code",
          "password"
        ],
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "user_id": "2",
          "code": "750248",
          "password": "newpassword"
        }
      },
...
      "NewAddress-Alreadyexist": {
        "title": "NewAddress-Alreadyexist",
        "required": [
          "message",
          "success"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "example": {
          "message": "Address already exist",
          "success": false
        }
      },
      "AddProductRequest": {
        "title": "AddProductRequest",
        "required": [
          "retail_price",
          "name",
          "description",
          "pic_filename",
          "stock"
        ],
        "type": "object",
        "properties": {
          "retail_price": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "pic_filename": {
            "type": "string"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "retail_price": 135.0,
          "name": "Blackpods Pro2",
          "description": "This is a scam",
          "pic_filename": "scampods_3529963a5-a2fa-458c-8b71-b4f8acdc1e12.jpg",
          "stock": 30
        }
      },
...
    }
  },
...
}